run hello world

+++ src/nes/cpu/opecode.rs
@@ -5,7 +5,7 @@ use std::{collections::HashMap};
 pub struct Opecode {
     pub code: Code,
     pub mode: AddressingMode,
-    pub cycle: u8,
+    pub cycle: u16,
 }
 
 #[derive(Debug, PartialEq)]
@@ -40,7 +40,7 @@ pub enum Code {
 }
 
 pub static OPECODE_MAP: Lazy<HashMap<u8, Opecode>> = Lazy::new(|| {
-    let cycles: Vec<u8> =
+    let cycles: Vec<u16> =
     vec![7, 6, 2, 8, 3, 3, 5, 5, 3, 2, 2, 2, 4, 4, 6, 6, 2, 5, 2, 8, 4, 4, 6, 6, 2, 4, 2, 7,
             4, 4, 6, 7, 6, 6, 2, 8, 3, 3, 5, 5, 4, 2, 2, 2, 4, 4, 6, 6, 2, 5, 2, 8, 4, 4, 6, 6,
             2, 4, 2, 7, 4, 4, 6, 7, 6, 6, 2, 8, 3, 3, 5, 5, 3, 2, 2, 2, 3, 4, 6, 6, 2, 5, 2, 8,
